Add files via upload
EncryptThis_CMD.jar is the command-line version of the program.  This still provides the user the option to run the GUI but the jar is only runnable from the command line.  

EncryptThis_UI.jar is the GUI version of the program.  This jar is executable by double-clicking the jar file in windows or from the command-line.  This version does not accept command-line arguments.